+++ src/index.d.ts
@@ -9,6 +9,7 @@ import {
 	getAwardStatistics,
 	getBadgeFields,
 	getCompletedAwards,
+	getCompletedAwardsByUser,
 	getContentAwards,
 	getContentAwardsByIds,
 	getInProgressAwards,
@@ -56,6 +57,7 @@ import {
 	getEnrichedLearningPath,
 	getEnrichedLearningPaths,
 	getLearningPathLessonsByIds,
+	isNextLessonLocked,
 	mapContentToParent,
 	resetAllLearningPaths,
 	startLearningPath,
@@ -175,7 +177,8 @@ import {
 	likePost,
 	search,
 	unlikePost,
-	updatePost
+	updatePost,
+	whoLikedPost
 } from './services/forums/posts.ts';
 
 import {
@@ -362,6 +365,10 @@ import {
 } from './services/sanity.js';
 
 import {
+	searchAlgolia
+} from './services/search.ts';
+
+import {
 	clearState
 } from './services/state.ts';
 
@@ -421,6 +428,7 @@ import {
 	fetchMemberships,
 	fetchRechargeTokens,
 	getUpgradePrice,
+	grant30DaysAccessForLifetime,
 	restorePurchases,
 	upgradeSubscription
 } from './services/user/memberships.ts';
@@ -454,9 +462,15 @@ import {
 } from './services/user/payments.ts';
 
 import {
+	fetchPlayerSettings,
+	updatePlayerSettings
+} from './services/user/playerSettings.ts';
+
+import {
 	deleteProfilePicture,
-	otherStats
-} from './services/user/profile.js';
+	otherStats,
+	updateProfileVisibility
+} from './services/user/profile.ts';
 
 import {
 	generateAuthSessionUrl,
@@ -486,6 +500,11 @@ import {
 	updatePracticeNotes,
 	updateUserPractice
 } from './services/userActivity.js';
+
+import {
+	whoLikedComment,
+	whoLikedContent
+} from './services/whoLiked.ts';
 
 import {
 	 default as EventsAPI 
@@ -611,6 +630,7 @@ declare module 'musora-content-services' {
 		fetchPackData,
 		fetchParentChildRelationshipsFor,
 		fetchPlayAlongsCount,
+		fetchPlayerSettings,
 		fetchPlaylist,
 		fetchPlaylistItems,
 		fetchPost,
@@ -665,6 +685,7 @@ declare module 'musora-content-services' {
 		getAwardStatistics,
 		getBadgeFields,
 		getCompletedAwards,
+		getCompletedAwardsByUser,
 		getContentAwards,
 		getContentAwardsByIds,
 		getContentRows,
@@ -722,6 +743,7 @@ declare module 'musora-content-services' {
 		getUserWeeklyStats,
 		getWeekNumber,
 		globalConfig,
+		grant30DaysAccessForLifetime,
 		guidedCourses,
 		hasAnyMethodV2IntroCompleted,
 		initializeEnvVar,
@@ -731,6 +753,7 @@ declare module 'musora-content-services' {
 		isContentLiked,
 		isContentLikedByIds,
 		isNextDay,
+		isNextLessonLocked,
 		isSameDate,
 		isUserFreeTier,
 		isUsernameAvailable,
@@ -790,6 +813,7 @@ declare module 'musora-content-services' {
 		restoreUserActivity,
 		restoreUserPractice,
 		search,
+		searchAlgolia,
 		sendAccountSetupEmail,
 		sendPasswordResetEmail,
 		setStudentViewForUser,
@@ -825,9 +849,11 @@ declare module 'musora-content-services' {
 		updateMultiUserAccount,
 		updateNotificationSetting,
 		updateOnboarding,
+		updatePlayerSettings,
 		updatePlaylist,
 		updatePost,
 		updatePracticeNotes,
+		updateProfileVisibility,
 		updateThread,
 		updateUserPractice,
 		upgradeSubscription,
@@ -836,6 +862,9 @@ declare module 'musora-content-services' {
 		userOnboardingForBrand,
 		verifyImageSRC,
 		verifyLocalDataContext,
+		whoLikedComment,
+		whoLikedContent,
+		whoLikedPost,
 	}
 }