+++ src/services/progress/internal/queries.ts
@@ -1,21 +1,18 @@
 import { db } from '../../sync'
-import ContentProgress, { CollectionParameter } from '../../sync/models/ContentProgress'
-import type { ModelSerialized } from '../../sync/serializers'
-
-type Selector<V> = (p: ModelSerialized<ContentProgress>) => V | null | undefined
+import { CollectionParameter } from '../../sync/models/ContentProgress'
 
 export const getByIds = async <V>(
   contentIds: number[],
-  select: Selector<V>,
-  fallback: V,
+  dataKey: string,
+  defaultValue: V,
   collection?: CollectionParameter
 ): Promise<Map<number, V>> => {
   if (contentIds.length === 0) return new Map()
 
-  const progress = new Map(contentIds.map((id) => [id, fallback]))
+  const progress = new Map(contentIds.map((id) => [id, defaultValue]))
   await db.contentProgress.getSomeProgressByContentIds(contentIds, collection).then((r) => {
     r.data.forEach((p) => {
-      progress.set(p.content_id, select(p) ?? fallback)
+      progress.set(p.content_id, p[dataKey] ?? defaultValue)
     })
   })
   return progress
@@ -23,26 +20,22 @@ export const getByIds = async <V>(
 
 export const getById = async <V>(
   contentId: number,
-  select: Selector<V>,
-  fallback: V,
+  dataKey: string,
+  defaultValue: V,
   collection?: CollectionParameter
 ): Promise<V> => {
-  if (!contentId) return fallback
+  if (!contentId) return defaultValue
   return db.contentProgress
     .getOneProgressByContentId(contentId, collection)
-    .then((r) => (r.data ? select(r.data) ?? fallback : fallback))
+    .then((r) => r.data?.[dataKey] ?? defaultValue)
 }
 
-export const getByRecordIds = async <V>(
-  ids: string[],
-  select: Selector<V>,
-  fallback: V
-): Promise<Record<string, V>> => {
-  const progress = Object.fromEntries(ids.map((id) => [id, fallback]))
+export const getByRecordIds = async <V>(ids: string[], dataKey: string, defaultValue: V) => {
+  const progress = Object.fromEntries(ids.map((id) => [id, defaultValue]))
 
   await db.contentProgress.getSomeProgressByRecordIds(ids).then((r) => {
     r.data.forEach((p) => {
-      progress[p.id] = select(p) ?? fallback
+      progress[p.id] = p[dataKey] ?? defaultValue
     })
   })